refactor(import-data): PR B move commit - git mv snapshot components to src/importdata
Pure file moves plus the package clause; nothing else changes and this
commit does not compile on its own. The adapt commit that follows adds the
per-component config structs and replaces the cmd globals with fields.

Moved: importDataState.go, importDataProgressReporter.go,
importDataFileTaskPicker.go, importDataSequentialFileBatchProducer.go,
importDataRandomFileBatchProducer.go, importDataFileTaskImporter.go,
importDataTestUtils.go and their five test files.
